+++ verifact/__init__.py
@@ -1,5 +1,5 @@
 from .invoice import Invoice
-from .app import app
+from .run_app import run
 from .error import ErrorWindow, run_error
 
-__all__ = ["Invoice", "app", "run_error"]
+__all__ = ["Invoice", "run", "run_error"]